+++ _includes/footer.html
@@ -1,26 +1,32 @@
 <footer class="site-footer h-card">
-    <data class="u-url" href="{{ " "="" |="" relative_url="" }}"=""></data>
+    <data
+        class="u-url"
+        href="{{ '/' | relative_url }}"
+    ></data>
     <div class="wrapper">
-      <div class="footer-col-wrapper">
-        <div class="footer-col footer-col-1">
-          <ul class="contact-list">
-            <li class="p-name">
-              {%- if site.author -%}
-                {{ site.author | escape }}
-              {%- else -%}
-                {{ site.title | escape }}
-              {%- endif -%}
-              </li>
-              {%- if site.email -%}
-              <li><a class="u-email" href="mailto:{{ site.email }}">{{ site.email }}</a></li>
-              {%- endif -%}
-          </ul>
-          {%- include social.html -%}
+        <div class="footer-col-wrapper">
+            <div class="footer-col footer-col-1">
+                <ul class="contact-list">
+                    <li class="p-name">
+                        {%- if site.author -%} {{ site.author | escape }} {%-
+                        else -%} {{ site.title | escape }} {%- endif -%}
+                    </li>
+                    {%- if site.email -%}
+                    <li>
+                        <a
+                            class="u-email"
+                            href="mailto:{{ site.email }}"
+                            >{{ site.email }}</a
+                        >
+                    </li>
+                    {%- endif -%}
+                </ul>
+                {%- include social.html -%}
+            </div>
+            <div class="footer-col footer-col-2"></div>
+            <div class="footer-col footer-col-3">
+                <p>{{- site.description | escape -}}</p>
+            </div>
         </div>
-        <div class="footer-col footer-col-2"></div>
-        <div class="footer-col footer-col-3">
-          <p>{{- site.description | escape -}}</p>
-        </div>
-      </div>
     </div>
-  </footer>
+</footer>
