+++ _includes/header.html
@@ -2,8 +2,8 @@
     <div class="wrapper">
       {%- assign default_paths = site.pages | map: "path" -%}
       {%- assign page_paths = site.header_pages | default: default_paths -%}
-      <a class="site-title" rel="author" href="{{ " "="" |="" relative_url="" }}"="">{{ site.title | escape }}</a>
-  
+      <a class="site-title" rel="author" href="{{ '/' | relative_url }}">{{ site.title | escape }}</a>
+
       {%- if page_paths -%}
         <nav class="site-nav">
           <input type="checkbox" id="nav-trigger" class="nav-trigger">
