+++ assets/js/pages/about.js
@@ -1,6 +1,4 @@
 import handleCookieLogic from '../terminal-loading-animation/helpers/handleCookieLogic.js';
 import terminalLoadingAnimation from '../terminal-loading-animation/terminalLoadingAnimation.js';
 
-
-
-window.addEventListener('load', handleCookieLogic(terminalLoadingAnimation));
+window.addEventListener('load', terminalLoadingAnimation);
