+++ assets/js/terminal-loading-animation/terminalLoadingAnimation.js
@@ -2,7 +2,6 @@ import setWrapperHeight from './helpers/setWrapperHeight.js';
 import createNodeArray from './helpers/createNodeArray.js';
 import wrapNodesInSpanTagAndApplyTypingClass from './helpers/wrapNodesInSpanTagAndApplyTypingClass.js';
 import animateElements from './helpers/animateElements.js';
-import handleCookieLogic from './helpers/handleCookieLogic.js';
 
 const terminalLoadingAnimation = () => {
     const nodes = [];
